Hurd, share uv_uptime and uv_loadavg with Linux
